change price field to decimal field in books model

+++ books/models.py
@@ -26,7 +26,7 @@ class Book(models.Model):
     author_id = models.ManyToManyField(
         Author, related_name="books", blank=True)
     title = models.CharField(max_length=250, unique=True)
-    price = models.FloatField()
+    price = models.DecimalField(max_digits=10, decimal_places=2)
     quantity = models.IntegerField()
     category_id = models.ForeignKey(
         Category, on_delete=models.CASCADE, related_name='books')
